requirements.txt
requests>=2.28,<3.0
google-genai>=1.0,<2.0
python-dotenv>=1.0,<2.0
python-dateutil>=2.8,<3.0
pyyaml>=6.0,<7.0
trafilatura>=1.6,<2.0
